+++ CHANGELOG.md
@@ -19,3 +19,4 @@
 - [2026-05-17 10:51] SKILL.md
 - [2026-05-17 11:06] home/godmode/config/prefill.json
 - [2026-05-17 11:08] home/godmode/lib/parsers.py
+- [2026-05-17 11:10] home/godmode/lib/encoder.py